Refactor: UI 모듈에 필요한 클래스 이동
