--- app/src/main/java/com/kappzzang/jeongsan/ui/Member.kt
@@ -1,3 +1,0 @@
-package com.kappzzang.jeongsan.ui
-
-class Member(val name: String)
