+++ ui/creategroup/src/main/java/com/kappzzang/jeongsan/creategroup/Member.kt
@@ -1,0 +1,3 @@
+package com.kappzzang.jeongsan.creategroup
+
+class Member(val name: String)
